# HBASE-28927 Fix spotbugs issues introduced by refactoring to hbase-di…