--- a/hbase-diagnostics/src/main/java/org/apache/hadoop/hbase/PerformanceEvaluation.java
+++ b/hbase-diagnostics/src/main/java/org/apache/hadoop/hbase/PerformanceEvaluation.java
@@ -260,6 +260,11 @@ public class PerformanceEvaluation extends Configured implements Tool {
       }
       return this.compareTo((RunResult) obj) == 0;
     }
+
+    @Override
+    public int hashCode() {
+      return Long.hashCode(duration);
+    }
   }
 
   /**