--- a/hbase-server/src/main/java/org/apache/hadoop/hbase/util/LoadTestKVGenerator.java
+++ b/hbase-server/src/main/java/org/apache/hadoop/hbase/util/LoadTestKVGenerator.java
@@ -33,6 +33,7 @@ public class LoadTestKVGenerator {
 
   private static final Logger LOG = LoggerFactory.getLogger(LoadTestKVGenerator.class);
   private static int logLimit = 10;
+  private static final Random SHARED_RANDOM = new Random();
 
   /** A random number generator for determining value size */
   private Random randomForValueSize = new Random(); // Seed may be set with Random#setSeed
@@ -98,7 +99,7 @@ public class LoadTestKVGenerator {
    * Generates random bytes of the given size for the given row and column qualifier. The random
    * seed is fully determined by these parameters.
    */
-  private static byte[] getValueForRowColumn(int dataSize, byte[]... seedStrings) {
+  private synchronized static byte[] getValueForRowColumn(int dataSize, byte[]... seedStrings) {
     long seed = dataSize;
     for (byte[] str : seedStrings) {
       final String bytesString = Bytes.toString(str);
@@ -106,9 +107,9 @@ public class LoadTestKVGenerator {
         seed += bytesString.hashCode();
       }
     }
-    Random seededRandom = new Random(seed);
+    SHARED_RANDOM.setSeed(seed);
     byte[] randomBytes = new byte[dataSize];
-    seededRandom.nextBytes(randomBytes);
+    SHARED_RANDOM.nextBytes(randomBytes);
     return randomBytes;
   }