@Apache-HBase 2024-10-22
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime |  Logfile | Comment |
|:----:|----------:|--------:|:--------:|:-------:|
| +0 :ok: |  reexec  |   0m 26s |  |  Docker mode activated.  |
|||| _ Prechecks _ |
| +1 :green_heart: |  dupname  |   0m  0s |  |  No case conflicting files found.  |
| +0 :ok: |  codespell  |   0m  0s |  |  codespell was not available.  |
| +0 :ok: |  detsecrets  |   0m  0s |  |  detect-secrets was not available.  |
| +1 :green_heart: |  @author  |   0m  0s |  |  The patch does not contain any @author tags.  |
| +1 :green_heart: |  hbaseanti  |   0m  0s |  |  Patch does not have any anti-patterns.  |
|||| _ master Compile Tests _ |
| +0 :ok: |  mvndep  |   0m 10s |  |  Maven dependency ordering for branch  |
| +1 :green_heart: |  mvninstall  |   2m 54s |  |  master passed  |
| +1 :green_heart: |  compile  |   3m 17s |  |  master passed  |
| +1 :green_heart: |  checkstyle  |   0m 43s |  |  master passed  |
| -1 :x: |  spotbugs  |   0m 27s | [/branch-spotbugs-hbase-diagnostics-warnings.html](https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/branch-spotbugs-hbase-diagnostics-warnings.html) |  hbase-diagnostics in master has 12 extant spotbugs warnings.  |
| -1 :x: |  spotbugs  |   1m 28s | [/branch-spotbugs-hbase-server-warnings.html](https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/branch-spotbugs-hbase-server-warnings.html) |  hbase-server in master has 1 extant spotbugs warnings.  |
| +1 :green_heart: |  spotless  |   0m 42s |  |  branch has no errors when running spotless:check.  |
|||| _ Patch Compile Tests _ |
| +0 :ok: |  mvndep  |   0m 11s |  |  Maven dependency ordering for patch  |
| +1 :green_heart: |  mvninstall  |   2m 44s |  |  the patch passed  |
| -1 :x: |  compile  |   0m 19s | [/patch-compile-hbase-diagnostics.txt](https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/patch-compile-hbase-diagnostics.txt) |  hbase-diagnostics in the patch failed.  |
| -0 :warning: |  javac  |   0m 19s | [/patch-compile-hbase-diagnostics.txt](https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/patch-compile-hbase-diagnostics.txt) |  hbase-diagnostics in the patch failed.  |
| +1 :green_heart: |  blanks  |   0m  0s |  |  The patch has no blanks issues.  |
| -0 :warning: |  checkstyle  |   0m  8s | [/results-checkstyle-hbase-diagnostics.txt](https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/results-checkstyle-hbase-diagnostics.txt) |  hbase-diagnostics: The patch generated 1 new + 43 unchanged - 1 fixed = 44 total (was 44)  |
| +1 :green_heart: |  spotbugs  |   1m 35s |  |  hbase-server generated 0 new + 0 unchanged - 1 fixed = 0 total (was 1)  |
| -1 :x: |  spotbugs  |   0m 36s | [/new-spotbugs-hbase-diagnostics.html](https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/new-spotbugs-hbase-diagnostics.html) |  hbase-diagnostics generated 1 new + 1 unchanged - 11 fixed = 2 total (was 12)  |
| +1 :green_heart: |  hadoopcheck  |  10m 38s |  |  Patch does not cause any errors with Hadoop 3.3.6 3.4.0.  |
| +1 :green_heart: |  spotless  |   0m 41s |  |  patch has no errors when running spotless:check.  |
|||| _ Other Tests _ |
| +1 :green_heart: |  asflicense  |   0m 17s |  |  The patch does not generate ASF License warnings.  |
|  |   |  37m 37s |  |  |


| Reason | Tests |
|-------:|:------|
| SpotBugs | module:hbase-diagnostics |
|  |  org.apache.hadoop.hbase.PerformanceEvaluation$RunResult defines equals and uses Object.hashCode()  At PerformanceEvaluation.java:Object.hashCode()  At PerformanceEvaluation.java:[lines 255-261] |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | ClientAPI=1.43 ServerAPI=1.43 base: https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/artifact/yetus-general-check/output/Dockerfile |
| GITHUB PR | https://github.com/apache/hbase/pull/6384 |
| Optional Tests | dupname asflicense javac spotbugs checkstyle codespell detsecrets compile hadoopcheck hbaseanti spotless |
| uname | Linux 9dc006bbba1c 5.4.0-1103-aws #111~18.04.1-Ubuntu SMP Tue May 23 20:04:10 UTC 2023 x86_64 x86_64 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | dev-support/hbase-personality.sh |
| git revision | master / d0d18ff89f6a9242627badeadf40cadaff14a7a2 |
| Default Java | Eclipse Adoptium-17.0.11+9 |
| Max. process+thread count | 84 (vs. ulimit of 30000) |
| modules | C: hbase-server hbase-diagnostics U: . |
| Console output | https://ci-hbase.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-6384/1/console |
| versions | git=2.34.1 maven=3.9.8 spotbugs=4.7.3 |
| Powered by | Apache Yetus 0.15.0 https://yetus.apache.org |


This message was automatically generated.